Simplify integer rank return in Rocket League calculator

+++ checkTierRL.py
@@ -65,7 +65,7 @@ class CheckTier(commands.Cog):
 
     def calculate_custom_league_rank(self, peak1, peak2):
         league_rank = max(peak1, peak2 - 120) * 0.75 + max(peak2, peak1 - 120) * 0.25
-        return int(round(league_rank))
+        return round(league_rank)
 
     def calculate_league_rank_1v1(self, peak3s, peak2s, peak1s):
         league_rank_1v1 = (0.2 * max(peak3s, peak2s - 120)) + (0.25 * max(peak2s, peak3s - 120)) + (0.8 * peak1s)
